+++ src/main/kotlin/com/github/libliboom/customplugintemplate/template/klass/Interactor.kt
@@ -2,24 +2,46 @@ package com.github.libliboom.customplugintemplate.template.klass
 
 import com.github.libliboom.customplugintemplate.manager.PackageManager
 
-fun createInteractorInput(
+fun createInteractor(
     interactorPackageName: String = PackageManager.interactorPackageName,
     className: String
 ) = """
   package $interactorPackageName
   
-  interface ${className}interactorInput{
+  class ${className}Interactor{
   
   }
 """.trimIndent()
 
-fun createInteractorOutput(
+fun createInteractorInterface(
     interactorPackageName: String = PackageManager.interactorPackageName,
     className: String
 ) = """
   package $interactorPackageName
   
-  interface ${className}interactorOutput{
+  interface ${className}InteractorInterface{
+  
+  }
+""".trimIndent()
+
+fun createInteractorDependenciesInterface(
+    interactorPackageName: String = PackageManager.interactorPackageName,
+    className: String
+) = """
+  package $interactorPackageName
+  
+  interface ${className}InteractorDependenciesInterface{
+  
+  }
+""".trimIndent()
+
+fun createInteractorModel(
+    interactorModelPackageName: String = PackageManager.interactorModelPackageName,
+    className: String
+) = """
+  package $interactorModelPackageName
+  
+  interface ${className}ItemInterface{
   
   }
 """.trimIndent()