+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 # Custom-Plugin-Template Changelog
 
-## [Unreleased]
+## 1.0.0
 ### Added
+- Add Dependencies interface to interactor
+### Fix
+- Fix execution bugs
+## 0.0.0
+### Added
+- Add businessLogic package
+- Add presenter and view
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)